restructure and simplify routes folder
