+++ src/routes/index.js
@@ -1,5 +1,13 @@
 var router = require('express').Router();
 
-router.use('/api', require('./api'));
+/* GET home page. */
+router.get('/', function(req, res, next) {
+  res.render('index', { title: 'Express' });
+});
+
+/* use next middleware functions */
+router.use('/users', require('./users'));
+router.use('/workouts', require('./workouts'));
+router.use('/diets', require('./diets'));
 
 module.exports = router;
